finish refactor SortingAlgorithm
