+++ src/test/java/pl/paweln/codility/counting/MissingIntegerTest.java
@@ -8,24 +8,24 @@ import pl.paweln.codility.core.*;
 public class MissingIntegerTest {
     private final SolutionFactory solutionFactory = new SolutionFactory();
 
-    private CodilitySolution s;
+    private CodilitySolution codilitySolution;
     private SolutionInputParams.SolutionInputParamsBuilder paramsBuilder;
     private SolutionInputParams params;
 
     @Before
-    public void setUp() {
-        this.s = this.solutionFactory.getSolution(SolutionFactory.CodilityTask.COUNTING_MISSING_INTEGER);
+    public void setup() {
+        this.codilitySolution = this.solutionFactory.getSolution(SolutionFactory.CodilityTask.COUNTING_MISSING_INTEGER);
         this.paramsBuilder = new SolutionInputParams.SolutionInputParamsBuilder();
     }
 
     @Test
-    public void testSolutionExample() {
+    public void testExample() {
         int[] testArr = new int[] {1, 3, 6, 4, 1, 2};
 
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
 
-        Assert.assertEquals(5, this.s.solution(this.params)[0]);
+        Assert.assertEquals(5, this.codilitySolution.solution(this.params)[0]);
     }
 
     @Test
@@ -35,7 +35,7 @@ public class MissingIntegerTest {
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
 
-        Assert.assertEquals(1, this.s.solution(this.params)[0]);
+        Assert.assertEquals(1, this.codilitySolution.solution(this.params)[0]);
     }
 
     @Test
@@ -45,7 +45,7 @@ public class MissingIntegerTest {
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
 
-        Assert.assertEquals(1, this.s.solution(this.params)[0]);
+        Assert.assertEquals(1, this.codilitySolution.solution(this.params)[0]);
 
     }
 
@@ -54,17 +54,17 @@ public class MissingIntegerTest {
         int[] testArr = new int[] {0};
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
-        Assert.assertEquals(1, this.s.solution(this.params)[0]);
+        Assert.assertEquals(1, this.codilitySolution.solution(this.params)[0]);
 
         testArr = new int[] {2};
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
-        Assert.assertEquals(1, this.s.solution(this.params)[0]);
+        Assert.assertEquals(1, this.codilitySolution.solution(this.params)[0]);
 
         testArr = new int[] {1};
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
-        Assert.assertEquals(2, this.s.solution(this.params)[0]);
+        Assert.assertEquals(2, this.codilitySolution.solution(this.params)[0]);
     }
 
     @Test
@@ -75,7 +75,7 @@ public class MissingIntegerTest {
         }
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
-        Assert.assertEquals(100000, this.s.solution(this.params)[0]);
+        Assert.assertEquals(100000, this.codilitySolution.solution(this.params)[0]);
     }
 
     @Test
@@ -83,11 +83,11 @@ public class MissingIntegerTest {
         int[] testArr = new int[] {0, 1, 2, 3, 4, 5, 6, 7};
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
-        Assert.assertEquals(8, this.s.solution(this.params)[0]);
+        Assert.assertEquals(8, this.codilitySolution.solution(this.params)[0]);
 
         testArr = new int[] {2, 3, 4, 5, 6, 7};
         this.params = this.paramsBuilder
                 .setFirstArray(testArr).build();
-        Assert.assertEquals(1, this.s.solution(this.params)[0]);
+        Assert.assertEquals(1, this.codilitySolution.solution(this.params)[0]);
     }
 }
